backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
google-genai
google-adk
pydantic
httpx
python-multipart
